LayoutTests/http/tests/referrer-policy-img/no-referrer/cross-origin-http-UpgradeMixedContent.https-expected.txt

Tests the behavior of no-referrer referrer policy when cross origin from HTTPS to HTTP, but the request is upgraded to HTTPS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is ""
PASS successfullyParsed is true

TEST COMPLETE
